+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+- Differentiate timeout errors (logged as `debug`) from other errors when performing mDNS lookup
+
 ## [0.20.0] - 2025-07-21
 
 - Optimize logging implementation by avoiding string allocating and formatting for messages that are below the current verbosity